extensions: preserve v0.8.0 ABI by introducing v0.9.0 WIT world
Add a new versioned WIT world (since_v0.9.0) for the workspace command
exports instead of mutating since_v0.8.0. This preserves the ABI for
extensions compiled against the existing 0.8.0 world.

since_v0.8.0 is refactored to a middle-version module that maps all
stable interfaces to the latest bindings and delegates host trait
implementations to since_v0_9_0 via the shared type mappings.
