+++ crates/extension_api/wit/since_v0.9.0/workspace-command.wit
